+++ crates/jsonschema-py/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.41.0] - 2026-02-04
+
 ### Changed
 
 - Update `pyo3` to `0.28`.
@@ -886,7 +888,8 @@
 ## 0.1.0 - 2020-06-09
 - Initial public release
 
-[Unreleased]: https://github.com/Stranger6667/jsonschema/compare/python-v0.40.4...HEAD
+[Unreleased]: https://github.com/Stranger6667/jsonschema/compare/python-v0.41.0...HEAD
+[0.41.0]: https://github.com/Stranger6667/jsonschema/compare/python-v0.40.4...python-v0.41.0
 [0.40.4]: https://github.com/Stranger6667/jsonschema/compare/python-v0.40.3...python-v0.40.4
 [0.40.3]: https://github.com/Stranger6667/jsonschema/compare/python-v0.40.2...python-v0.40.3
 [0.40.2]: https://github.com/Stranger6667/jsonschema/compare/python-v0.40.1...python-v0.40.2